requirements.txt
-c constraints.txt
django
djangorestframework
drf-extensions
pyyaml
uritemplate
